refactor: context & reducer
