+++ src/index.tsx
@@ -4,7 +4,7 @@ import './tailwind.output.css'
 import App from './App';
 import { BrowserRouter as Router} from "react-router-dom";
 import reportWebVitals from './reportWebVitals';
-import { QuizProvider } from "./dataProvider/QuizProvider";
+import { QuizProvider } from "./dataProvider/context/Quiz/QuizProvider";
 
 ReactDOM.render(
   <React.StrictMode>
