+++ src/pages/ShowQuestions.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useParams } from "react-router";
-import { useQuiz } from "../dataProvider/QuizProvider";
+import { useQuiz } from "../dataProvider/context/Quiz/QuizProvider";
 import { Link } from "react-router-dom";
 import firebase from "../firebase/firebase";
 import "firebase/firestore";
